+++ webapi/src/test/scala/org/knora/webapi/store/triplestore/TestDatasetBuilder.scala
@@ -15,14 +15,13 @@ import zio.TaskLayer
 import zio.UIO
 import zio.ZIO
 import zio.ZLayer
-
 import java.io.StringReader
 
 object TestDatasetBuilder {
 
-  def readToModel(turtle: String)(model: Model): Model = model.read(new StringReader(turtle), null, "TTL")
+  private def readToModel(turtle: String)(model: Model): Model = model.read(new StringReader(turtle), null, "TTL")
 
-  def transactionalWrite(change: Model => Model)(ds: Dataset): Task[Dataset] = ZIO.attempt {
+  private def transactionalWrite(change: Model => Model)(ds: Dataset): Task[Dataset] = ZIO.attempt {
     ds.begin(ReadWrite.WRITE)
     try {
       change apply ds.getDefaultModel
@@ -33,12 +32,12 @@ object TestDatasetBuilder {
     ds
   }
 
-  val createTxnMemDataset: UIO[Dataset] = ZIO.succeed(DatasetFactory.createTxnMem())
+  private val createTxnMemDataset: UIO[Dataset] = ZIO.succeed(DatasetFactory.createTxnMem())
 
-  def datasetFromTurtle(turtle: String): Task[Dataset] =
+  private def datasetFromTurtle(turtle: String): Task[Dataset] =
     createTxnMemDataset.flatMap(transactionalWrite(readToModel(turtle)))
 
-  def asLayer(ds: Task[Dataset]): TaskLayer[Ref[Dataset]] = ZLayer.fromZIO(ds.flatMap(Ref.make[Dataset](_)))
+  private def asLayer(ds: Task[Dataset]): TaskLayer[Ref[Dataset]] = ZLayer.fromZIO(ds.flatMap(Ref.make[Dataset](_)))
 
   def datasetLayerFromTurtle(turtle: String): TaskLayer[Ref[Dataset]] = asLayer(datasetFromTurtle(turtle))
   val emptyDataSet: TaskLayer[Ref[Dataset]]                           = datasetLayerFromTurtle("")
